[Windows] Move HID console output to separate window
