+++ windows/QMK Toolbox/Usb/Bootloader/LufaHidDevice.cs
@@ -1,4 +1,4 @@
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
 
 namespace QMK_Toolbox.Usb.Bootloader
 {
